t/100-symlinker: Extract setup_test_dirs()
This creates the standard build and test dirs (the test dir is
guaranteed to be empty) and sets up varaibles for them.

This is intended later to be extracted out to a common code file.

Potential bug: the method of finding the base dir for the checkout
works only for test files directly under the `t/` subdir.