symlinker: Handle conflicts
Conflicts may be "in-home" (with other files under ~/.home/ or
created by the setup script) or "out-home" (files from elsewhere,
created before the setup script ran).

In both cases we leave the first file that was there alone and
print a warning about subsequent files that would conflict.

This means that "out-home" files take priority (and get left
alone). For inside files, non-dangling links created before the
setup script ran are preserved, otherwise the highest priority
(from a lexical sort of `~/.home/*/bin/*`) link will be created
as usual and further attempted links to a file of the same name
will generate a conflict message.