t/100-symlinker: Always show failure for both diff and output
We diff the file tree as before, displaying the differences, but
now do this before we check the output of the program. However,
we delay asserting the failure of the diff (if it failed) until
after asserting the program output was correct. This means that
if both fail, we now see the actual/expected for both rather than
just the program output.