symlinker: Function to find dest link/dir and link target from src
Tweak test framework and program to allow for unit testing
functions in the `bin/dot-home-setup` script.

Add function that, given a source file or directory dir under
`~/.home` calculates where a link or directory goes under `$HOME`
and, if it's a link, what the relative link back to `~/.home`
would be. This makes somewhat tricky use of Bash parameter
expansion and pattern substitution to count the number of path
components minus one by counting the number of slashes.

This will be used in future commits.