Test: Remove unneeded empty dir cleanup
This appears to have been in place because checking out previous
revisions can leave behind empty directories that would then cause the
test to fail. Previous revisions of this Test script will still take
care of this, but at this point we no longer list directories in the
expected or actual output, so we don't need this.