Test: Pull up setting of build dir vars and cleaning build test dir
Set `$base_dir`, `$build_dir` and `$build_t_dir` (this last one for
data generated by tests) in the environment at the start of the test
run so that they're available to everything in the test framework (and
build framework, too, if we ever need to build anything).

At the start of the test run we ensure that `$build_t_dir` exists but
is empty. This ensures it happens only once at the start of a test
run, rather than risking tests wiping out other tests' results.

Now use `.build/t` rather than `.build/test` to hold data generated by
tests to match the directory for test code and source data.