t/100-symlinker: Better test scratch dir and vars setup
In setup() we now set variables for test suite and test name, a source
path for the suite (so we can find data files that go along with the
test) and a scratch dir for the particular test.

The scratch dir is under `$build_t_dir` and named for the test suite
and test.

Our test name is different from $BATS_TEST_NAME in that we don't have
a (redundant) `test_` prefix on the front of it. However, it will still
have the Bats substituations for special characters, turning `a_b` into
`a-2d5fb` for example.