symlinker: Don't use readlink -f
`readlink -f` fails with no output if the link is dangling.
Thus if a file in `.home` wants a link at that spot, and a
dangling link is already there, we don't remove it and put
in the proper link.

Also, `readlink` doesn't have a `-f` option on MacOS X (even
though `stat` does).

Also tweak a comment for another test.