dot-home-setup: Start adding inb4 pass (t/2xx tests)
The inb4 pass doesn't actually do anything yet; this is simply
the framework to call and test it.

We add a `-p` option to `dot-home-setup` to specify which pass(es) to
call. (This option can be specified multiple times; passes will be
called in the order specified.) If not given, all passes are called
(in the correct order) by default.

Add the t/t260-inb4-functional test and data; this currently just
verifies the input without generating any output.

Next steps:
* Create a _built_ file with the contents of the source files
* Check to see that the _built_ file is copied to the _installed_ file
* Check that it won't rebuild if the _installed_ file is different from
  the _built_ file from the previous run