symlinker: Don't create ~/dot/
When we removed the trailing slash from `dot/` in the find command in
83254f6 (to fix MacOS find adding an extra slash in the path), this
exposed a bug in set_dest_target() where it would not replace `dot`
with `.` if there was no trailing slash.

We can't easily match "slash or end of string" in the Bash pattern
subtitution (it uses shell globs, not regular expressions) so instead
we handle this as a special case.

Now with `dot` properly being turned into `.` again, the `mkdir -p .`
is harmless.