test-lib: New "test homedir" test framework
This gives us functions to up a new, empty "home" dir with specific
files and links in it, run `dot-home-setup` on that dir, and check the
resulting file tree, nicely printing any differences.

We split the printing of the diff and the assertion that there was no
difference so that we can do other assertions in between. This means
that when `assert_ouptut` fails we can see both the ouptut diff and
the file tree diff.