110-symlinker: Convert 100-symlinker tests to new framework
We can now use many small test cases instead of one huge one; we
also break down the tests in a slightly different way. This is
unfortunately now longer (65 lines → 165 lines overall) and
arguably harder to read as a whole, but the individual tests are
now much easier to understand and the various bits of behaviour
are much better isolated. Perhaps we can find a clever way to get
rid of some of the additional boilerplate we've introduced by
splitting one enormous test into many small ones.

There are additional failing (but skipped) tests added as well
that demonstrate some bugs we need to fix.