t/100-symlinker: Remove remaining stub of old functional tests
The functional tests (which actually run the symnlinker on a directory
tree) is now covered by the new t/110-symlinker tests. (The unit test
for `set_dest_target()` remains in place.)

We also leave in place the (minimal) `setup_test_suite_home` framework
in `t/test-lib.bash` until we can confirm we don't need it for tests
of other things we're adding soon.