dot-home-setup: Better and tested behaviour when ~/.home nonexistent
We send the error message from `cd` to `/dev/null` because we generate
our own error message after it fails.

Unskip the test for this, and add a test for `~/.home` existing but
having no modules (but an ignored dir). This already worked ok,
silently doing nothing, though perhaps we want to think about printing
a warning for that.