Test: Add -3 option to test with Bash 3 instead of system Bash
This is a start on being able to test with alternative tools as used
on non-GNU/Linux systems (BSD and MacOS).

If given the -3 option we download and build Bash 3, install it into
`.build/local` and prefix `.build/local/bin` to our path which should
(theoretially) make it used everywhere in the tests. Other tools can
be added to `.build/local` later.

We have a separate `.archive/` directory, rather than putting it under
`.build`, so that forcing a clean build with `rm -rf .build` doesn't
cause another download.

MacOS is the only commonly used system we know of these days that's
still using Bash 3, but that's unlikely to change because Apple's not
upgraded to Bash 4 due to licensing issues.