Extract functions from set_dest_target
Given lists of source files, we need to work out the "destinations"
(i.e., where we place the symlink that links to the source file) and
the "targets" (i.e., the relative paths those symlinks use to point
back to the source file).

set_dest_target() previously did all of this together with special
cases to determine if the file was a "bin" or "dot" file; instead now
we have separate add_bin_dest() and add_dot_dest() to work out the
destination for files from under `bin/` and `dot/`, respectively, and
symlink_rel_target() to handle the relative paths for any given dest.

We also make use of the newly introduced "spec" or "specification"
test system to more easily and clearly unit test symlink_rel_target().