t/test-lib: Extract and expand trim_spec function
This trims and decomments what we're now calling a "specification,"
which is a set of lines where leading/trailing spaces, comments and
blank lines are ignored. (Further interpretation is left to whatever's
reading the spec.)

This and related stuff is likely to be pulled out into a separate lib
at some point.