assert_function: Don't let function under test read spec input
The FUT is running inside a `while read spec` loop that's reading from
stdin, so if the function also reads anything from stdin it will be
consuming the specification input that should be going to the
enclosing loop. Not only does this break the tests, but also produces
very confusing output.