Implement fetching of dependencies
The dependency fetching is completely orthogonal to the updates
pass; only modules existing at the start of the `dot-home-setup`
run are updated as newly-fetched modules are already up to date.

Unlike updates, the dependency-fetching pass is always run because
it's a reasonably efficient no-op when all dependencies are present
and the user's dot-home configuration is likely to be fairly broken
if dependencies are missing.

Each module may have a `dh/dependencies` file listing dependences
as (non-blank, non-comment) lines with whitespace-separated:
  * module name
  * dependency type (currently only "git" is supported)
  * other parameters (the path to the repo to clone for "git")

To handle dependencies in newly fetched dependencies we keep
re-running the fetch pass until no new modules appear. This handles
both the case where everything was fetched and the case where a fetch
failed. This does have the annoying side-effect of attempting to fetch
failed dependencies every time `dot-home-setup` is run; at the moment
the user must fix this by fixing the fetch error or working around it
by, e.g., creating an empty module.

The tests include bare repos to be fetched as fixtures. Though
all the refs are packed in these repos, we still must have a
`refs` subdir for Git to recognise these as repos.