t/test-lib: Add assert_success_and_diff_test_home_with
We were never actually checking that run_setup_on_test_home didn't
exit with an error code. Part of the reason was probably that this had
to be done as a separate line in every test, either an
`assert_success` or `assert_failure` as appropriate.

We introduce a new assert_success_and_diff_test_home_with to combine
the two without adding an extra line in the most common case.