t/110-symlinker: Check for conflicts in aliased directories
An aliased directory is one where the same file can be installed from
more than one location in a single module, e.g., `bin/X` as an
alternative for `dot/local/bin/X`.

Really we ought to print out all the files that are conflicting with
each other (and perhaps install none of them), rather than installing
the first one and then printing out only subsequent files that
conflict, but that's work for a later time.