Add setup pass to run */dh/setup scripts
We run the dh/setup scripts (executables of any kind with that name,
actually) after all the other passes to ensure that they're working
with a completely updated tree. Otherwise they might not see things
that they'd use that have not yet come in.

The CWD for these scripts is defined to be in the module itself, and
we do a couple of simple checks for this in the tests.

The testing of this isn't really complete; it's entirely in the "All
together now" pass and doesn't test that we get a proper warning about
error code returns from trying to run the script (though we do
actually print it and it's been tested "by hand").