Standard updater now aborts rebase if it failed
A failed rebase will leave behind the conflicts which are likely to
leave broken files in production since we carry on to rebuild/relink
those files after that. So if it fails, abort the rebase, leaving us
where we were before with the previously working files. The user can
go attempt to fix it up manually later.

We hide error messages from `--abort` because the rebase may have not
even started for other reasons (e.g., presence of unstaged changes),
but the 'No rebase in progress?' message in that case is more
confusing than useful.