Test: Hack #include for BSD sed build
In the build of the BSD sed from the Apple sources, fgetln() in main.c
was returning an invalid pointer. This was due to #include'ing the GNU
<stdio.h> rather than <bsd/stdio.h>, which apparently worked at some
point in the past but is now incompatible.

So we hack around this by modifying the stdio.h #include to instead
include the BSD library version.

We also tried building instead with `-I/usr/include/bsd`, but that
results in:

    /usr/include/bsd/sys/cdefs.h:30:23: error: #include nested too deeply

and a slew of other errors, so it appears we need to use the BSD
header files on a case-by-case basis.