Test: Don't build Bash 3 and BSD sed if we're on MacOS
Instead we can just use the native tools, since they're already what
we want.  We still require the -3 option to be passed in, however,
because we want always to make it clear when we're testing against
Bash-3/BSD-sed (and to remind MacOS devs that they need to run their
tests on a Bash-4/GNU-sed system as well).

Reviewed-by: Kevin Kee <kevinkee9@gmail.com>