t/test-lib: Fix variable interpolation for MacOS Bash bug
On MacOS 10.14 a variable interpolation seems to consider part of the
UTF-8 for a `∙` character trailing a variable name to be part of the
variable name itself, despite LC_ALL being (apparently) set correctly.
This produced some 'bad encoding' messages from sed, which is why we
believe it may have been only part of the UTF-8 sequence being
misinterpreted this way.

Fixed by using braces to force the `∙` not to be considered part of
the variable name.

This may be a bug in Bash; the version we build has a patch as
compared to the MacOS version (see below). Or it may be some other
library-related thing.

Mac:   GNU bash, version 3.2.57(1)-release (x86_64-apple-darwin18)
Linux: GNU bash, version 3.2.57(2)-release (x86_64-unknown-linux-gnu)

Reviewed-by: Kevin Kee <kevinkee9@gmail.com>