Test,t/010-framework: Build and use BSD head(1) command
We had inadvertently added `head -q` to dot-home-setup, which fails on
MacOS because the BSD head(1) doesn't support this option. (The option
suppresses the extra header lines showing the filename when multiple
files are given to head(1).)

Now we build the BSD head(1), along with sed(1), and confirm that
we're using in the tests. This now causes a failure with `Test -3` (in
the "inb4build pass" and integration tests), demonstrating how it was
failing on MacOS.