Test,t/010-framework: Bash version is either 3 or ≥4, not just =4
Bash 5 has been released (and is now shipping on Arch Linux). It
doesn't seem to have significant incompatibilities, at least for our
purposes, so we continue to test we work on v3 and whatever modern
version of Bash the developer is using, which should continue to cover
all versions under which we run.

Reviewed-by: Curt J. Sampson <cjs@cynic.net>