.bashrc.inb1: Add prepath() and prepath $HOME/.local/bin
$HOME/.local/bin is an opinionated path into which dot-home installs things
from modules' bin/ directory. We define this in order to allow easily and
safely adding that to the path, as not all systems do this by default.

This will now, unlike before, produce a conflict notice for .bashrc for new
users that have just installed dot-home. This is is fine because a) this
shows that dot-home runs safely (avoiding destroying any user data), b)
introduces users to this feature immediately, and c) .bashrc is generally
the first thing people will be moving to dot-home management anyway.

The documentation needs to be updated to describe this.