.bashrc.inb1: Override prepath() def; mkdir .local/bin; add docs
We override any existing prepath() definition because we use it
immediately after defining it and expect it to work in a certain way.
This isn't an undue burden on users because if they really need to
redefine it, they can do it in a dot/*.inb3 file.

Create the ~/.local/bin/ directory if it doesn't already exist. It
should already exist anyway, but just in case this prevents an error
from prepath().

Add some documentation about our usage of ~/.local/.