bootstrap-user: Handle no entry in $DH_BOOTSTRAP_USERS and existing repos
If there's no entry for $LOGNAME in the file referenced by
$DH_BOOTSTRAP_USERS we print a warning and carry on rather than
printing and error and halting. This makes life easier for new users,
as they can now use this script too.

We also now handle the situation where the user repo or dot-home/
already exists (or the directory, at least, exists) by simply fetching
and rebasing it, so that this script can be re-run if necessary. (The
fetch and rebase may or may not work, depending on what's really in
that directory, of course.)