for_each_module(): Allow periods in module names; just check first char
Previously we did not allow module names with periods in them to
prevent directories such as `fred.git` from being treated as modules.
But this breaks otherwise sensible module names such as
`mycompany.com`.

Change our ignore rules so that anything that starts with a letter
(upper or lower case) or number is considered a module, and only
files/directories starting with punctuation (or non-ASCII letters) are
ignored.

This means that we can no longer put `local.git` (as a remote for
`local`) under .home/, though it's not clear why we would be putting
that there anyway, but it can be worked around by calling it
`,local.git`.